Move neo4j inserter into insta/inserter/neo4j
