+++ insta/inserter/neo4j/posts/main.go
@@ -2,21 +2,21 @@ package main
 
 import (
 	"encoding/json"
-	"github.com/codeuniversity/smag-mvp/neo4j-inserter"
 
 	"github.com/codeuniversity/smag-mvp/kafka"
 	"github.com/codeuniversity/smag-mvp/kafka/changestream"
+	neo4jinserter "github.com/codeuniversity/smag-mvp/neo4j/inserter"
 	"github.com/codeuniversity/smag-mvp/service"
 	"github.com/codeuniversity/smag-mvp/utils"
+
 	"github.com/neo4j/neo4j-go-driver/neo4j"
 )
 
 func main() {
-
 	readerConfig := kafka.GetInserterConfig()
 	neo4jConfig := utils.GetNeo4jConfig()
 
-	i := neo4jinserter.New(neo4jConfig, kafka.NewReader(readerConfig), addTaggedUsersRelationship)
+	i := neo4jinserter.New(neo4jConfig, kafka.NewReader(readerConfig), insertPostsAndAddRelationship)
 
 	service.CloseOnSignal(i)
 	waitUntilClosed := i.Start()
@@ -24,25 +24,25 @@ func main() {
 	waitUntilClosed()
 }
 
-type taggedUser struct {
+type Post struct {
 	UserID int `json:"user_id"`
-	PostID int `json:"post_id"`
+	PostID int `json:"id"`
 }
 
-func addTaggedUsersRelationship(m *changestream.ChangeMessage, session neo4j.Session) error {
-	const addTaggedRelationship = `
+func insertPostsAndAddRelationship(m *changestream.ChangeMessage, session neo4j.Session) error {
+	const insertPostsAndAddRelationship = `
 	MERGE(u:USER{id: $userID})
 	MERGE(p:POST{id: $postID})
-	MERGE(u)-[:TAGGED_ON]->(p)
+	MERGE(u)-[:POSTED]->(p)
 	`
-	t := &taggedUser{}
-	err := json.Unmarshal(m.Payload.After, t)
+	p := &Post{}
+	err := json.Unmarshal(m.Payload.After, p)
 
 	if err != nil {
 		return err
 	}
 
-	_, err = session.Run(addTaggedRelationship, map[string]interface{}{"userID": t.UserID, "postID": t.PostID})
+	_, err = session.Run(insertPostsAndAddRelationship, map[string]interface{}{"userID": p.UserID, "postID": p.PostID})
 
 	if err != nil {
 		return err
