+++ neo4j/inserter/inserter.go
@@ -10,8 +10,8 @@ import (
 	"github.com/codeuniversity/smag-mvp/kafka/changestream"
 	"github.com/codeuniversity/smag-mvp/utils"
 	"github.com/codeuniversity/smag-mvp/worker"
-	"github.com/neo4j/neo4j-go-driver/neo4j"
 
+	"github.com/neo4j/neo4j-go-driver/neo4j"
 	kf "github.com/segmentio/kafka-go"
 )
